extracted icon buton into separate file
