+++ packages/ui-components/src/components/button/button.tsx
@@ -27,55 +27,6 @@ export const SimpleButton: React.FC<ButtonProps> = ({
   );
 };
 
-// ICON BUTTON ================================================================
-
-export type IconButtonProps = ButtonProps & {
-  /** Wheter the icon is left or right of the label */
-  side: 'left' | 'right';
-  /** The icon's source */
-  iconSrc: string;
-};
-
-/** Button with settable icon. The icon can be specified via its source and can
- * be but left or right of the lable. */
-export const IconButton: React.FC<IconButtonProps> = ({
-  mode = 'primary',
-  size = 'default',
-  side = 'left',
-  iconSrc,
-  label,
-  onClick,
-  disabled,
-}) => {
-  return (
-    <StyledButton mode={mode} size={size} onClick={onClick} disabled={disabled}>
-      <FlexDiv side={side}>
-        <StyledIcon iconSrc={iconSrc} />
-        <p>{label}</p>
-      </FlexDiv>
-    </StyledButton>
-  );
-};
-
-type FlexDivProps = {
-  side: IconButtonProps['side'];
-};
-
-const FlexDiv = styled.div.attrs(({side}: FlexDivProps) => {
-  let className = 'flex space-x-2';
-  if (side === 'left') className = 'flex space-x-2';
-  else className = 'flex flex-row-reverse space-x-reverse space-x-2';
-  return {className: className};
-})<FlexDivProps>``;
-
-type StyledIconProps = {
-  iconSrc: IconButtonProps['iconSrc'];
-};
-
-const StyledIcon = styled.img.attrs(({iconSrc}: StyledIconProps) => {
-  return {src: iconSrc, className: 'w-3 h-3'};
-})<StyledIconProps>``;
-
 // Auxiliary Components ========================================================
 
 type SizedButtonProps = {
@@ -100,22 +51,24 @@ type StyledButtonProps = {
 /**
  * Extends the SizedButton element with the desired styling
  */
-const StyledButton = styled(SizedButton).attrs(({mode}: StyledButtonProps) => {
-  let className;
+export const StyledButton = styled(SizedButton).attrs(
+  ({mode}: StyledButtonProps) => {
+    let className;
 
-  if (mode === 'primary') {
-    className =
-      'text-ui-0 bg-primary-400 hover:bg-primary-500 active:bg-primary-700';
-  } else if (mode === 'secondary') {
-    className =
-      'text-primary-500 bg-primary-100 hover:text-primary-800 active:bg-primary-200 active:text-primary-800';
-  } else if (mode === 'tertiary') {
-    className =
-      'text-ui-600 bg-ui-0 border-2 border-ui-100 hover:border-ui-300 active:border-ui-800';
-  } else if (mode === 'ghost') {
-    className =
-      'text-primary-500 bg-ui-0 hover:text-primary-800 active:bg-primary-50';
+    if (mode === 'primary') {
+      className =
+        'text-ui-0 bg-primary-400 hover:bg-primary-500 active:bg-primary-700';
+    } else if (mode === 'secondary') {
+      className =
+        'text-primary-500 bg-primary-100 hover:text-primary-800 active:bg-primary-200 active:text-primary-800';
+    } else if (mode === 'tertiary') {
+      className =
+        'text-ui-600 bg-ui-0 border-2 border-ui-100 hover:border-ui-300 active:border-ui-800';
+    } else if (mode === 'ghost') {
+      className =
+        'text-primary-500 bg-ui-0 hover:text-primary-800 active:bg-primary-50';
+    }
+
+    return {className};
   }
-
-  return {className};
-})<StyledButtonProps>``;
+)<StyledButtonProps>``;
