+++ packages/ui-components/src/components/button/index.tsx
@@ -1,3 +1,4 @@
-export * from './button';
+export {ButtonProps, SimpleButton} from './button';
 export * from './addButton';
 export * from './openButton';
+export * from './iconButton';
